refactor: 바뀐 타입 적용 #37
